Add parsevariable util
